+++ packages/jsts/src/rules/S5256/rule.ts
@@ -24,7 +24,7 @@ import * as estree from 'estree';
 import { Rule } from 'eslint';
 import { TSESTree } from '@typescript-eslint/utils';
 import { getProp, getLiteralPropValue } from 'jsx-ast-utils';
-import { isPresentationTable } from '../helpers';
+import { computeGrid, isPresentationTable } from '../helpers';
 import { isHtmlElement } from '../helpers/isHtmlElement';
 
 export const rule: Rule.RuleModule = {
@@ -74,147 +74,3 @@ export const rule: Rule.RuleModule = {
     };
   },
 };
-
-type TableCell = {
-  rowSpan: number;
-  isHeader: boolean;
-};
-
-function computeSpan(tree: TSESTree.JSXElement, spanKey: string): number {
-  let span = 1;
-  const spanAttr = getProp(tree.openingElement.attributes, spanKey);
-  if (spanAttr) {
-    span = parseInt(String(getLiteralPropValue(spanAttr)));
-  }
-  return span;
-}
-
-function rowSpan(tree: TSESTree.JSXElement): number {
-  let value = computeSpan(tree, 'rowspan');
-  const MAX_ROW_SPAN = 65534;
-  if (value > MAX_ROW_SPAN) {
-    value = MAX_ROW_SPAN;
-  }
-  return value;
-}
-
-function colSpan(tree: TSESTree.JSXElement): number {
-  let value = computeSpan(tree, 'colspan');
-  const MAX_INVALID_COL_SPAN = 10000;
-  if (value > MAX_INVALID_COL_SPAN) {
-    value = 1;
-  }
-  return value;
-}
-
-function extractRow(tree: TSESTree.JSXElement): TableCell[] {
-  const row: TableCell[] = [];
-  tree.children.forEach(child => {
-    if (child.type !== 'JSXElement') {
-      return;
-    }
-    const colSpanValue = colSpan(child);
-    const rowSpanValue = rowSpan(child);
-    for (let i = 0; i < colSpanValue; i++) {
-      row.push({
-        rowSpan: rowSpanValue,
-        isHeader:
-          child.openingElement.name.type === 'JSXIdentifier' &&
-          child.openingElement.name.name === 'th',
-      });
-    }
-  });
-  return row;
-}
-
-function extractRows(context: Rule.RuleContext, tree: TSESTree.JSXElement): TableCell[][] | null {
-  const rows: TableCell[][] = [];
-  let unknownTableStructure = false;
-  tree.children.forEach(child => {
-    if (child.type === 'JSXElement') {
-      const childType = getElementType(context)(child.openingElement).toLowerCase();
-      if (childType === 'tr') {
-        rows.push(extractRow(child));
-      } else if (childType === 'table') {
-        // skip
-      } else {
-        const KNOWN_TABLE_STRUCTURE_ELEMENTS = ['thead', 'tbody', 'tfoot'];
-        if (KNOWN_TABLE_STRUCTURE_ELEMENTS.includes(childType)) {
-          const extractedRows = extractRows(context, child);
-          if (extractedRows === null) {
-            unknownTableStructure = true;
-          } else if (extractedRows.length > 0) {
-            rows.push(...extractedRows);
-          }
-        } else if (!isHtmlElement(child)) {
-          unknownTableStructure = true;
-        }
-      }
-    } else if (
-      child.type === 'JSXExpressionContainer' &&
-      child.expression.type !== 'JSXEmptyExpression'
-    ) {
-      unknownTableStructure = true;
-    }
-  });
-  if (unknownTableStructure) {
-    return null;
-  }
-  return rows;
-}
-
-function computeGrid(context: Rule.RuleContext, tree: TSESTree.JSXElement): boolean[][] | null {
-  const rows = extractRows(context, tree);
-  if (rows === null) {
-    return null;
-  }
-  if (rows.length === 0) {
-    return [];
-  }
-  const nbColumns = rows[0].length;
-  const columns: (TableCell | undefined)[] = Array.from({ length: nbColumns });
-  let row = 0;
-  const result = [];
-  while (row < rows.length) {
-    const resultRow = [];
-    let rowIndex = 0;
-    // Checks if any of the cells in the current row that is added was used from the incoming rows[row]
-    let usedCurrentRow = false;
-    // Checks if row was built entirely out of columns with rowSpan == 0
-    let onlyMaxRowSpan = true;
-    for (let column = 0; column < nbColumns; column++) {
-      if (!columns[column]) {
-        if (rowIndex === rows[row].length) {
-          // We have reached the end of the current row from the table definition
-          continue;
-        }
-        columns[column] = rows[row][rowIndex];
-        rowIndex++;
-        usedCurrentRow = true;
-      }
-      const currentCell = columns[column];
-      if (!currentCell) {
-        continue;
-      }
-      resultRow.push(currentCell.isHeader);
-      if (currentCell.rowSpan > 0) {
-        // Mark that there is at least one cell that is not built entirely out of columns with rowSpan == 0
-        onlyMaxRowSpan = false;
-        currentCell.rowSpan--;
-        if (currentCell.rowSpan === 0) {
-          columns[column] = undefined;
-        }
-      }
-    }
-    if (onlyMaxRowSpan) {
-      // If the row was built entirely out of columns with rowSpan == 0, we finish the construction
-      break;
-    }
-    result.push(resultRow);
-    if (usedCurrentRow) {
-      // Increment the row index only if we used any of the cells from the incoming rows[row]
-      row++;
-    }
-  }
-  return result;
-}
